Nahod 18 dev/anton250

+++ src/base/models.py
@@ -1113,7 +1113,7 @@ class Employment(AbstractActiveModel):
 
     # new worker restrictions
     week_availability = models.SmallIntegerField(default=7)
-    norm_work_hours = models.SmallIntegerField(default=100)
+    norm_work_hours = models.FloatField(default=100)
     shift_hours_length_min = models.SmallIntegerField(blank=True, null=True)
     shift_hours_length_max = models.SmallIntegerField(blank=True, null=True)
     min_time_btw_shifts = models.SmallIntegerField(blank=True, null=True)
